chore: FavoriteViewSet - Authentication

+++ api/routes/favorite/views.py
@@ -4,6 +4,7 @@ from rest_framework.permissions import IsAuthenticated
 
 from api.models import Favorite
 from api.serializers import FavoriteSerializer
+from authentication.models import ProfileUser
 from utils.jsend_responses import *
 
 
@@ -13,25 +14,37 @@ class FavoriteViewSet(viewsets.ModelViewSet):
     permission_classes = [IsAuthenticated]
 
     def list(self, request):
-        favorites = Favorite.objects.all()
+        favorites = Favorite.objects.filter(owner=request.user)
         serializer = FavoriteSerializer(favorites, many=True)
         return success_response(key="favorites", data=serializer.data)
 
     def retrieve(self, request, pk=None):
         try:
             favorite = self.get_object()
-            serializer = FavoriteSerializer(favorite)
-            return success_response(key="favorite", data=serializer.data)
-        except Exception:
+
+            if favorite.owner == request.user:
+                serializer = FavoriteSerializer(favorite)
+                return success_response(key="favorite", data=serializer.data)
+
+            return fail_response(
+                errors={"user": "You are not Favorite Owner"},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
+        except Http404:
             return error_response("Favorite not found")
 
     def create(self, request):
         serializer = FavoriteSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
-            return success_response(
-                {"favorite": serializer.data}, status=status.HTTP_201_CREATED
+            user = serializer.validated_data.get("owner")
+            if user != None and user == request.user:
+                serializer.save()
+                return success_response(key="favorite", data=serializer.data)
+            return fail_response(
+                {"user": "You are not Favorite Owzner"},
+                status=status.HTTP_401_UNAUTHORIZED,
             )
+
         return fail_response(serializer.errors)
 
     def update(self, request, pk=None):
@@ -41,15 +54,28 @@ class FavoriteViewSet(viewsets.ModelViewSet):
             return error_response("Favorite not found")
 
         serializer = FavoriteSerializer(favorite, data=request.data)
+
         if serializer.is_valid():
-            serializer.save()
-            return success_response(serializer.data)
+            if favorite.owner == request.user:
+                serializer.save()
+                return success_response(key="favorite", data=serializer.data)
+
+            return fail_response(
+                {"user": "You are not Favorite Owner"},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         return fail_response(serializer.errors)
 
     def destroy(self, request, pk=None):
         try:
             favorite = self.get_object()
-            favorite.delete()
+            if favorite.owner == request.user:
+                favorite.delete()
+                return success_response()
+
+            return fail_response(
+                {"user": "You are not Favorite Owner"},
+                status=status.HTTP_401_UNAUTHORIZED,
+            )
         except Http404:
             return error_response("Favorite not found")
-        return success_response()
